refactor: 정적 파일 못 찾을시 404 페이지 응답하도록 변경
- StaticFileLoader Application으로 이동
