+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -1,10 +1,10 @@
 package com.techcourse.controller;
 
+import com.techcourse.controller.util.StaticFileLoader;
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
 import java.util.Optional;
 import org.apache.coyote.http11.ContentType;
-import org.apache.coyote.http11.StaticFileLoader;
 import org.apache.coyote.http11.controller.AbstractController;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
