+++ tomcat/src/main/java/com/techcourse/controller/util/StaticFileLoader.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package com.techcourse.controller.util;
 
 import java.io.File;
 import java.io.IOException;
