+++ tomcat/src/main/java/org/apache/coyote/http11/response/ResponseStatus.java
@@ -6,6 +6,7 @@ public enum ResponseStatus {
     FOUND(302, "Found"),
     UNAUTHORIZED(401, "Unauthorized"),
     BAD_REQUEST(400, "Bad Request"),
+    NOT_FOUND(404, "Not Found"),
     ;
 
     private final int code;
